requirements.txt
supabase
discord.py==2.3.2
python-dotenv
asynctest
pytest
arrow
psycopg2-binary
pandas
numpy
matplotlib
seaborn
clickhouse_connect